Allow to symbolicate 64-bit arch addresses

+++ natos/NATOS.h
@@ -15,9 +15,9 @@
 @property (nonatomic, copy, readwrite) NSString* XCArchivePath;
 @property (nonatomic, copy, readwrite) NSString* dSYMPath;      // if not present, derived from XCArchivePath
 @property (nonatomic, copy, readwrite) NSString* CPUArchitecture;
-@property (nonatomic, assign, readwrite) unsigned int mainFunctionStackAddress;
-@property (nonatomic, assign, readwrite) unsigned int loadAddress;
-@property (nonatomic, assign, readwrite) unsigned int targetStackAddress;
+@property (nonatomic, assign, readwrite) unsigned long long mainFunctionStackAddress;
+@property (nonatomic, assign, readwrite) unsigned long long loadAddress;
+@property (nonatomic, assign, readwrite) unsigned long long targetStackAddress;
 
 - (void) printUsage;
